requirements.txt
EAS2Text-Remastered
apprise
pytz
